Unify eth_simulateV1 types in x/evm/types
The SimOpts / SimBlock / SimBlockOverrides / SimCallResult /
SimBlockResult types were duplicated between rpc/types and x/evm/types.
The two copies were structurally identical, SimError was already
single-sourced on the keeper side, and TransactionArgs (used by
SimBlock.Calls) already crossed the boundary the same way, so the split
had no semantic meaning and only created drift risk (both files carried
"keep in sync with..." comments).

Make x/evm/types the single source of truth:

- Move UnmarshalJSON for SimBlockResult and MarshalJSON for
  SimCallResult onto the keeper-side types so the struct round-trips
  across the gRPC wire without a duplicate definition.
- Rewire rpc/backend, rpc/namespaces/ethereum/eth, and the
  EthereumAPI / EVMBackend interfaces to use evmtypes.SimOpts and
  evmtypes.SimBlockResult directly, matching the pattern
  TransactionArgs already follows.
- Drop rpc/types/simulate_v1.go and consolidate its unique test
  coverage into x/evm/types/simulate_v1_test.go.
