+++ rpc/namespaces/ethereum/eth/api.go
@@ -87,7 +87,7 @@ type EthereumAPI interface {
 	// Allows developers to read data from the blockchain which includes executing
 	// smart contracts. However, no data is published to the network.
 	Call(args evmtypes.TransactionArgs, blockNrOrHash rpctypes.BlockNumberOrHash, overrides *rpctypes.StateOverride) (hexutil.Bytes, error)
-	SimulateV1(opts rpctypes.SimOpts, blockNrOrHash *rpctypes.BlockNumberOrHash) ([]*rpctypes.SimBlockResult, error)
+	SimulateV1(opts evmtypes.SimOpts, blockNrOrHash *rpctypes.BlockNumberOrHash) ([]*evmtypes.SimBlockResult, error)
 
 	// Chain Information
 	//
